Refactor routine tests
